Modify commands to allow inspec suggest to be a top-level command
